VisionVend/raspberry_pi/requirements.txt
opencv-python-headless==4.7.0.72
transformers==4.33.3
torch==2.0.1
torchvision==0.15.2
RPi.GPIO==0.7.1
pyyaml==6.0.1
picamera2==0.3.13
